:art: refactor : Proccessor에 몰린 코드를 각 객체의 역할에 맞춰 옮기는 것으로 개선
