+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,20 +1,14 @@
 package org.apache.coyote.http11;
 
 import static org.apache.coyote.http11.common.Constants.CRLF;
-import static org.apache.coyote.http11.common.Constants.SPACE;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.common.QueryString;
 import org.apache.coyote.http11.common.RequestHeader;
 import org.apache.coyote.http11.common.RequestURI;
 import org.apache.coyote.http11.common.ResponseEntity;
@@ -24,13 +18,6 @@ import org.slf4j.LoggerFactory;
 public class Http11Processor implements Runnable, Processor {
 
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
-    private static final String LOGIN_PAGE = "/login.html";
-    private static final String INDEX_PAGE = "/index.html";
-    private static final String LOGIN_FAIL_PAGE = "/401.html";
-    private static final String QUERY_STRING_BEGIN_CHAR = "?";
-    private static final int HTTP_METHOD_INDEX = 0;
-    private static final int URI_INDEX = 1;
-    private static final int HTTP_VERSION_INDEX = 2;
 
     private final Socket connection;
 
@@ -53,7 +40,7 @@ public class Http11Processor implements Runnable, Processor {
             final var bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
             final var firstLine = bufferedReader.readLine();
 
-            final var requestURI = readURI(firstLine);
+            final var requestURI = RequestURI.from(firstLine, user -> log.info("user : {}", user));
             final var requestHeader = readHeader(bufferedReader);
 
             final var responseEntity = ResponseEntity.from(requestURI);
@@ -66,82 +53,12 @@ public class Http11Processor implements Runnable, Processor {
         }
     }
 
-    private RequestURI readURI(final String requestURILine) {
-        final var requestURIElements = parseRequestURIElements(requestURILine);
-        final var uri = requestURIElements.get(URI_INDEX);
-
-        if (uri.startsWith("/login")) {
-            return parseLoginRequestURI(uri, requestURIElements);
-        }
-
-        return parseRequestURI(uri, QueryString.empty(), requestURIElements);
-    }
-
-    private List<String> parseRequestURIElements(final String requestURILine) {
-        return Arrays.stream(requestURILine.split(SPACE))
-                .collect(Collectors.toUnmodifiableList());
-    }
-
-    private RequestURI parseLoginRequestURI(final String uri, final List<String> requestURIElements) {
-        final var index = uri.indexOf(QUERY_STRING_BEGIN_CHAR);
-
-        if (isExistQueryString(index)) {
-            final var queryString = QueryString.from(uri.substring(index + 1));
-
-            final var account = queryString.getValue("account");
-            final var password = queryString.getValue("password");
-            return InMemoryUserRepository.findByAccount(account)
-                    .filter(user -> user.checkPassword(password))
-                    .map(user -> loginSuccess(user, queryString, requestURIElements))
-                    .orElseGet(() -> loginFail(requestURIElements));
-        }
-
-        return parseRequestURI(LOGIN_PAGE, QueryString.empty(), requestURIElements);
-    }
-
-    private boolean isExistQueryString(final int index) {
-        return index != -1;
-    }
-
-    private RequestURI loginSuccess(
-            final User user,
-            final QueryString queryString,
-            final List<String> requestURIElements
-    ) {
-        log.info("user : {}", user);
-        return parseRequestURI(
-                INDEX_PAGE,
-                queryString,
-                requestURIElements
-        );
-    }
-
-    private RequestURI loginFail(final List<String> requestURIElements) {
-        return parseRequestURI(
-                LOGIN_FAIL_PAGE,
-                QueryString.empty(),
-                requestURIElements
-        );
-    }
-
-    private RequestURI parseRequestURI(
-            final String uri,
-            final QueryString queryString,
-            final List<String> requestURIElements
-    ) {
-        return RequestURI.of(
-                uri,
-                queryString,
-                requestURIElements.get(HTTP_METHOD_INDEX),
-                requestURIElements.get(HTTP_VERSION_INDEX)
-        );
-    }
-
     private RequestHeader readHeader(final BufferedReader bufferedReader) throws IOException {
         final StringBuilder stringBuilder = new StringBuilder();
         for (String line = bufferedReader.readLine(); !"".equals(line); line = bufferedReader.readLine()) {
             stringBuilder.append(line).append(CRLF);
         }
+
         return RequestHeader.from(stringBuilder.toString());
     }
